+++ projects/client/src/lib/features/portal/_internal/alignPopupContainer.ts
